web/requirements.txt
numpy
Pillow
ai-edge-litert
